Creating the Model Mongoose models are created from schemas and are responsible for creating and reading documents from the underlying MongoDB database.
Model Definition:
The code checks if a model named "users" already exists in the mongoose.models object. This object holds all registered models.
If the model exists, it reuses the existing model.
If the model does not exist, it creates a new model using the userSchema.
This is done to avoid recompiling the model if the code is executed multiple times (for example, in a development environment with hot-reloading).