Creating a Hashed Token Generating the Hashed Token:
const hashedToken = await bcryptjs.hash(userId.toString(), 10);
The userId is converted to a string and hashed using bcryptjs with a salt rounds value of 10. This creates a secure, hashed token.
Encoding the Token:

const encodedToken = encodeURIComponent(hashedToken);
The hashed token is URL-encoded to ensure it can be safely included in the URL without breaking.