Logging the Request Body:
console.log(reqBody);: Logs the request body to the console for debugging purposes.
Checking if the User Already Exists:

const user = await User.findOne({ email }): Checks if a user with the given email already exists in the database.
If a user is found, it returns a JSON response indicating the user already exists and sets the status code to 400 (Bad Request).