Sending a Verification Email:
await sendEmail({ email, emailType: "VERIFY", userId: savedUser._id }): Sends a verification email to the user using the sendEmail function. The email type is set to "VERIFY" and the userId is passed to identify the user.
The console logs (console.log(1) and console.log(2)) are used to track the progress of the email sending process.