Finding the User by Token:
const user = await User.findOne({ verifyToken: token, verifyTokenExpiry: { $gt: Date.now() } });
This query searches for a user whose verifyToken matches the provided token and whose verifyTokenExpiry is greater than the current time (i.e., the token is still valid).
If no user is found (i.e., the token is invalid or expired), a response is returned with an error message and a status code of 400 (Bad Request).