Configuration:
export const config = { ... }: Defines configuration options for the middleware.
matcher: Specifies an array of paths to which the middleware should be applied. In this case, it includes paths for the root, profile, login, signup, and verifyemail routes.
Summary:
This middleware function ensures that users are redirected appropriately based on their authentication status when accessing different routes in the Next.js application. It restricts access to private routes for unauthenticated users and redirects them to the login page. Similarly, it redirects authenticated users away from public routes like login, signup, and verifyemail to prevent redundant access.